+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-** Features **:
+**Features**:
+
+- Limit the number of custom measurements per event. ([#1483](https://github.com/getsentry/relay/pull/1483)))
 - Add INP web vital as a measurement. ([#1487](https://github.com/getsentry/relay/pull/1487))
 
 **Internal**: